% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Handywages.R
\name{Handywages_delete}
\alias{Handywages_delete}
\title{计时计件工资表删除}
\usage{
Handywages_delete(dms_token, FID)
}
\arguments{
\item{FID}{}
}
\value{

}
\description{
计时计件工资表删除
}
\examples{
Handywages_delete()
}
