% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Handywages.R
\name{Handywages_upload}
\alias{Handywages_upload}
\title{上传工资信息}
\usage{
Handywages_upload(
  dms_token,
  FYEAR,
  FMONTH,
  FNAME,
  FBaseWages,
  FattendanceDays,
  FleaveDays
)
}
\arguments{
\item{FleaveDays}{}
}
\value{

}
\description{
上传工资信息
}
\examples{
Handywages_upload()
}
