% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HourlyPiecerate.R
\name{HourlyPiecerate_delete}
\alias{HourlyPiecerate_delete}
\title{计时计件工资表删除}
\usage{
HourlyPiecerate_delete(token, a.FProductionNumber, Fseq)
}
\arguments{
\item{Fseq}{}
}
\value{

}
\description{
计时计件工资表删除
}
\examples{
HourlyPiecerate_delete()
}
