% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HourlyPiecerate.R
\name{HourlyPiecerate_detail_view}
\alias{HourlyPiecerate_detail_view}
\title{计时计件工资mingxi表查询}
\usage{
HourlyPiecerate_detail_view(dms_token, FYEAR, FMONTH)
}
\value{

}
\description{
计时计件工资mingxi表查询
}
\examples{
HourlyPiecerate_view()
}
