% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HourlyPiecerate.R
\name{HourlyPiecerate_preview}
\alias{HourlyPiecerate_preview}
\title{计时计件工资表预览}
\usage{
HourlyPiecerate_preview(token, file_name)
}
\arguments{
\item{token}{}
}
\value{

}
\description{
计时计件工资表预览
}
\examples{
HourlyPiecerate_preview()
}
