% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HourlyPiecerate.R
\name{HourlyPiecerate_upload}
\alias{HourlyPiecerate_upload}
\title{计时计件工资表上传}
\usage{
HourlyPiecerate_upload(token, file_name)
}
\arguments{
\item{token}{}
}
\value{

}
\description{
计时计件工资表上传
}
\examples{
HourlyPiecerate_upload()
}
