% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HourlyRouting.R
\name{HourlyRouting_delete}
\alias{HourlyRouting_delete}
\title{计时计件工资表删除}
\usage{
HourlyRouting_delete(token, FRoutingNumber)
}
\arguments{
\item{FRoutingNumber}{}
}
\value{

}
\description{
计时计件工资表删除
}
\examples{
HourlyRouting_delete()
}
