% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HourlyRouting.R
\name{HourlyRouting_upload}
\alias{HourlyRouting_upload}
\title{计时计件工资表上传}
\usage{
HourlyRouting_upload(token, file_name)
}
\arguments{
\item{token}{}
}
\value{

}
\description{
计时计件工资表上传
}
\examples{
HourlyRouting_upload()
}
