% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HourlyRouting.R
\name{HourlyRouting_view}
\alias{HourlyRouting_view}
\title{计时工资表查询}
\usage{
HourlyRouting_view(token, FRoutingNumber)
}
\arguments{
\item{FRoutingNumber}{}
}
\value{

}
\description{
计时工资表查询
}
\examples{
HourlyRouting_view()
}
