% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HourlyWages.R
\name{HourlyWages_delete}
\alias{HourlyWages_delete}
\title{计时计件工资表删除}
\usage{
HourlyWages_delete(token, FOPERATIONNAMEFSEQ)
}
\arguments{
\item{FOPERATIONNAMEFSEQ}{}
}
\value{

}
\description{
计时计件工资表删除
}
\examples{
HourlyWages_delete()
}
