% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HourlyWages.R
\name{HourlyWages_upload}
\alias{HourlyWages_upload}
\title{计时计件工资表上传}
\usage{
HourlyWages_upload(token, file_name)
}
\arguments{
\item{token}{}
}
\value{

}
\description{
计时计件工资表上传
}
\examples{
HourlyWages_upload()
}
