% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HourlyWages.R
\name{HourlyWages_view}
\alias{HourlyWages_view}
\title{计时工资表查询}
\usage{
HourlyWages_view(token, FOPERATIONNAME)
}
\arguments{
\item{FOPERATIONNAME}{}
}
\value{

}
\description{
计时工资表查询
}
\examples{
HourlyWages_view()
}
