% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/productprice.R
\name{productprice_delete}
\alias{productprice_delete}
\title{计时计件工资表删除}
\usage{
productprice_delete(token, FSpecification)
}
\arguments{
\item{FSpecification}{}
}
\value{

}
\description{
计时计件工资表删除
}
\examples{
productprice_delete()
}
