% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/productprice.R
\name{productprice_upload}
\alias{productprice_upload}
\title{计时计件工资表上传}
\usage{
productprice_upload(token, file_name)
}
\arguments{
\item{token}{}
}
\value{

}
\description{
计时计件工资表上传
}
\examples{
productprice_upload()
}
