% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/productprice.R
\name{productprice_view}
\alias{productprice_view}
\title{计时工资表查询}
\usage{
productprice_view(token, FSpecification)
}
\arguments{
\item{FRoutingNumber}{}
}
\value{

}
\description{
计时工资表查询
}
\examples{
productprice_view()
}
