% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/productRouting.R
\name{productRouting_delete}
\alias{productRouting_delete}
\title{计时计件工资表删除}
\usage{
productRouting_delete(token, FSpecification)
}
\arguments{
\item{FSpecification}{}
}
\value{

}
\description{
计时计件工资表删除
}
\examples{
productRouting_delete()
}
