% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/productRouting.R
\name{productRouting_upload}
\alias{productRouting_upload}
\title{计时计件工资表上传}
\usage{
productRouting_upload(token, file_name)
}
\arguments{
\item{token}{}
}
\value{

}
\description{
计时计件工资表上传
}
\examples{
productRouting_upload()
}
