% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/productRouting.R
\name{productRouting_view}
\alias{productRouting_view}
\title{计时工资表查询}
\usage{
productRouting_view(token, FSpecification)
}
\arguments{
\item{FRoutingNumber}{}
}
\value{

}
\description{
计时工资表查询
}
\examples{
productRouting_view()
}
